refactor: move server to infra/transport/http/server
