+++ solution/go.mod
@@ -3,6 +3,8 @@ module github.com/wdsjk/avito-shop
 go 1.23.6
 
 require (
+	github.com/go-chi/chi v1.5.5
+	github.com/go-chi/chi/v5 v5.2.3
 	github.com/ilyakaznacheev/cleanenv v1.5.0
 	github.com/jackc/pgx/v5 v5.7.5
 	github.com/lib/pq v1.10.9
